+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 
 This changelog starts with the restructured 1.0.0 release that happened after context isolation changes. The changelogs here should more-or-less mirror the ones that get shown in the client but probably with less formatting and pizzazz.
 
+1.13.7
+### Added
+- New Webpack cache system for faster loading times; you will see much faster loading times after the first launch with this update.
+
+### Fixed
+- Fixed `tags` button appearing behind other elements in the Addon Store
+- Fixed downloading addons via the addon store
+- Fixed dropdowns with lots of options not allowing you to scroll through all of the options
+- Minor UI fixes
+
+### Improvements
+- Dropdowns now automatically scroll to the selected option when opened
+
+1.13.6
+
+### Fixed
+- Fixed BetterDiscord not appearing in the `Settings Cog` context menu (right click menu)
+- Fixed `Version Info` strings in settings not appearing
+- Native titlebar matches Linux now
+
+### Improvements
+- Allow context menu toggle to be preventDefaulted
+- ContextMenu API Rework
+- Rewrite Native Fetch
+
+1.13.5
+
+### Added
+- Added `data-speaking` attribute to VoiceUser in ThemeAttributes
+
+### Fixed
+- Fixed BdApi.Net.fetch not handling relative urls properly
+- BetterDiscord now properly checks for addon updates periodically
+- BetterDiscord now loads up correctly
+
+### Improved
+- Tweaked styles for InstallCSS
+- BetterDiscord's Dropdown Component was rewrote
+
+### Removed
+
 1.13.4
 
 ### Fixed: